requirements.txt
python-dotenv
langchain
langchain-mistralai
langchain_community
langchain_huggingface
transformers
huggingface_hub
bitsandbytes
accelerate
